# specs/001-prd-app-only/contracts/api.yaml
openapi: 3.0.3
info:
  title: Yabaii - Japanese Price Comparison App API
  description: RESTful API for Japanese price comparison mobile application
  version: 1.0.0
  contact:
    name: API Team
    url: https://yabaii.day
  license:
    name: MIT
servers:
  - url: https://api.yabaii.day/v1
    description: Production
  - url: https://staging-api.yabaii.day/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /recommend/{spuId}:
    get:
      tags: [Recommendations]
      summary: Get product recommendations
      operationId: getRecommendations
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [similar, alternative, trending]
            default: similar
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Product recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationsResponse'

  # Community endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Recommendation:
      type: object
      properties:
        productId:
          type: string
        type:
          type: string
        score:
          type: number
        reason:
          type: string

    # Additional response schemas
...
    RecommendationsResponse:
      type: object
      properties:
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
...